backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
aiofiles
python-multipart
pydantic
